fix: #5 Entity, Dto 자료형 수정

+++ src/main/java/com/example/vegan_life/dto/MemberRequestDto.java
@@ -21,8 +21,8 @@ public class MemberRequestDto {
     private String newPassword;
     private String name;
     private String nickname;
-    private Float height;
-    private Float weight;
+    private Integer height;
+    private Integer weight;
     private String activationRatio;
     private VegeType vegeType;
     private LocalDateTime createdAt;
